requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-google-genai
langchain_core
langchainhub
langchain_experimental
psycopg2-binary